requirements.txt
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
